requirements.txt
click==8.1.3
colorama==0.4.5
distlib==0.3.6
filelock==3.8.0
Flask==2.2.2
Flask-Cors==3.0.10
Flask-SQLAlchemy==2.5.1
greenlet==1.1.3
gunicorn==20.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mysql-connector-python==8.0.30
platformdirs==2.5.2
psycopg2==2.9.3
PyMySQL==1.0.2
six==1.16.0
SQLAlchemy==1.4.41
virtualenv==20.16.5
Werkzeug==2.2.2
